Create Pentesting2
References


Official NMap Tutorial
Metasploit Databases
Hydra Github



Key Terms



Host Discovery: The process of discovering live hosts on a network.


Example: One way to do this is ping every IP address in the, subnet and see which ones respond. This is important because it enables attackers to identify potential targets.



Service Discovery: The process of discovering which services are running on a reachable host.


Example: The first step is typically a port scan, which reveals which ports are open on a machine.



Passive OS Detection: nmap guesses which operating system is running on the target of a port scan by looking for clues in a machine's responses.


Example: A TCP response generated by a Windows machine will contain the same raw data as one generated by a Linux machine, but the Time to Live will often be different: Windows sets it to 128 by default, while Linux has 64.



Active OS Detection: More accurate than nmap's passive detection, but also more invasive and more likely to be identified.


Example: You should only use Active OS Detection on one host at a time.



Brute Force Attack: Also known as a 'dictionary attack' An attacker uses a list of passwords to try to log in as a given user.


Example: An attacker might try to log in as the user admin with the passwords password; administrator; 12345; etc, hoping that one of them will work.


Pillaging the act of exfiltrating valuable data from a compromised host. This is the phase during which you collect sensitive information about your target.
Fuzzing: The act of sending fake data in order to "trick" the service into behaving a certain way. A brute-force attack against an SSH server using the ssh_login module is a good example.

Metasploit has modules for three different kinds of scanning:



Port Scans: Metasploit can perform TCP connect and SYN scans.

Service Scans: Metasploit has a wealth of modules for scanning specific services. These modules are useful for detecting the version of a specific service.
They can also be used for more advanced purposes, such as "fuzzing" the target service.

Post-Exploitation Enumerations: Scans that are run on compromised hosts after you've hacked into them. These can be used to scan for information like a list of installed applications, users, running processes/services, and/or stored credentials.


Post-exploitation is the most complicated phase of a pentest, because there are so many things an attacker might try to do. Some common examples include:
Gather data about the compromised host. One of the first things attackers do after compromising a host is learn as much as they can about it.


Enumeration is the process of of learning as much as you can about a victim machine, often from system and configuration files. Metasploit provides a wealth of modules specifically for post-exploitation data enumeration.


Maintain Access. If you exploit a host and lose your connection, the only way to get it back is to exploit it again. To avoid this, attackers often leave "backdoors" on compromised targets.

Steal data. Stealing data is commonly called data exfiltration or pillaging.    


Example: In addition to the information above, attackers will look for client-specific data, such as financial records; business documents; confidental communications; etc.


Pivot. Pentesters often manage to compromise machines that have multiple network interfaces. Each network interface attached to a machine allows it to communicate with a different subnet. This means that compromising a machine with multiple network interfaces might allow attackers to pivot into completely new networks. This sometimes results in their compromising machines on sensitive internal networks that were assumed to be safe, which can have disastrous consequences: Exfiltration of employee PII, confidential financial records, executive correspondences, etc.